Added Customer and Seller services.

+++ BackEnd/src/main/java/com/communitycart/BackEnd/service/FIleStorage.java
@@ -13,7 +13,7 @@ import java.nio.file.Paths;
 public class FIleStorage {
 
     public String saveImages(String loc, MultipartFile file, Long productId) throws Exception {
-        Path path = Paths.get(loc);
+        Path path = Paths.get("C:\\Program Files\\Apache Software Foundation\\Tomcat 10.1\\webapps\\images\\product");
         try{
             String fileName = file.getOriginalFilename();
             String ext = fileName.substring(fileName.indexOf('.'));
@@ -25,9 +25,60 @@ public class FIleStorage {
                     break;
                 }
             }
-            fileToBeDeleted.delete();
+            if(fileToBeDeleted != null){
+                fileToBeDeleted.delete();
+            }
             Files.copy(file.getInputStream(), path.resolve(productId.toString() + ext));
             return productId + ext;
+        } catch (Exception e) {
+            throw new Exception(e.getMessage());
+        }
+
+    }
+
+    public String saveSellerPhoto(MultipartFile file, Long sellerId) throws Exception {
+        Path path = Paths.get("C:\\Program Files\\Apache Software Foundation\\Tomcat 10.1\\webapps\\images\\shop");
+        try{
+            String fileName = file.getOriginalFilename();
+            String ext = fileName.substring(fileName.indexOf('.'));
+            File folder = new File(path.toString());
+            File fileToBeDeleted = null;
+            for(File fi: folder.listFiles()){
+                if(fi.getName().equals(sellerId + ext)){
+                    fileToBeDeleted = fi;
+                    break;
+                }
+            }
+            if(fileToBeDeleted != null){
+                fileToBeDeleted.delete();
+            }
+            Files.copy(file.getInputStream(), path.resolve(sellerId.toString() + ext));
+            return sellerId + ext;
+        } catch (Exception e) {
+            throw new Exception(e.getMessage());
+        }
+
+    }
+
+
+    public String saveCustomerPhoto(MultipartFile file, Long customerId) throws Exception {
+        Path path = Paths.get("C:\\Program Files\\Apache Software Foundation\\Tomcat 10.1\\webapps\\images\\customers");
+        try{
+            String fileName = file.getOriginalFilename();
+            String ext = fileName.substring(fileName.indexOf('.'));
+            File folder = new File(path.toString());
+            File fileToBeDeleted = null;
+            for(File fi: folder.listFiles()){
+                if(fi.getName().equals(customerId + ext)){
+                    fileToBeDeleted = fi;
+                    break;
+                }
+            }
+            if(fileToBeDeleted != null){
+                fileToBeDeleted.delete();
+            }
+            Files.copy(file.getInputStream(), path.resolve(customerId.toString() + ext));
+            return customerId + ext;
         } catch (Exception e) {
             throw new Exception(e.getMessage());
         }
